refactor(ReceiveAmountScreen): move and rename
